Feature/improved users table

+++ components/app/user_action_button.tsx
@@ -12,7 +12,12 @@ import { User } from "@supabase/supabase-js";
 import { Fragment, useState } from "react";
 import { usePopper } from "react-popper";
 import Swal from "sweetalert2";
+import EditUserDetails from "./EditUserDetails";
 import UserInfo from "./UserInfo";
+import { getUser } from "@/lib/supabase/client";
+import { redirect } from "next/navigation";
+import { useRouter } from "next/navigation";
+import { useOpenStore } from "@/store/useOpenStore";
 
 const jsonTheme = {
   main: "line-height:1.3;color:#383a42;background:#ffffff;overflow:hidden;word-wrap:break-word;white-space: pre-wrap;word-wrap: break-word; ",
@@ -28,9 +33,18 @@ function classNames(...classes: any[]) {
   return classes.filter(Boolean).join(" ");
 }
 
-export default function UserActionButton({ selectedUser }: { selectedUser: User }) {
+export default function UserActionButton({
+  selectedUser,
+  open,
+  setOpen,
+}: {
+  selectedUser: User;
+  open: boolean;
+  setOpen: any;
+}) {
   const [referenceElement, setReferenceElement] = useState(null);
   const [popperElement, setPopperElement] = useState(null);
+
   const { styles, attributes } = usePopper(referenceElement, popperElement, {
     placement: "bottom-end",
     modifiers: [
@@ -42,39 +56,49 @@ export default function UserActionButton({ selectedUser }: { selectedUser: User 
       },
     ],
   });
-  const deleteBtn = () => {
-    Swal.fire({
-      title: "Are you sure?",
-      text: "You won't be able to revert this!",
-      icon: "warning",
-      showCancelButton: true,
-      confirmButtonText: "Yes, delete it!",
-      cancelButtonText: "No, cancel!",
-      reverseButtons: true,
-    }).then(async (result) => {
-      if (result.isConfirmed) {
-        const response = await deleteUser(selectedUser.id);
+  const deleteBtn = async () => {
+    const { user } = await getUser();
 
-        if (!response.error) {
-          Swal.fire({
-            title: "Deleted!",
-            text: "The user successfully deleted.",
-            icon: "success",
-          }).then(() => {
-            location.reload(); // Reload the page
-          });
-        } else {
-          Swal.fire({
-            title: "Failed!",
-            text: response.error.message,
-            icon: "error",
-          });
+    if (selectedUser.id === user?.id) {
+      Swal.fire({
+        title: "Failed!",
+        text: "You can't delete your own account.",
+        icon: "error",
+      });
+    } else {
+      Swal.fire({
+        title: "Are you sure?",
+        text: "You won't be able to revert this!",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonText: "Yes, delete it!",
+        cancelButtonText: "No, cancel!",
+        reverseButtons: true,
+      }).then(async (result) => {
+        if (result.isConfirmed) {
+          const response = await deleteUser(selectedUser.id);
+
+          if (!response.error) {
+            Swal.fire({
+              title: "Deleted!",
+              text: "The user successfully deleted.",
+              icon: "success",
+            }).then(() => {
+              // redirect("./"); // Reload the page
+              location.reload();
+            });
+          } else {
+            Swal.fire({
+              title: "Failed!",
+              text: response.error.message,
+              icon: "error",
+            });
+          }
         }
-      }
-    });
+      });
+    }
   };
 
-  const [open, setOpen] = useState(false);
   return (
     <>
       <Menu
@@ -260,7 +284,7 @@ export default function UserActionButton({ selectedUser }: { selectedUser: User 
                       <div className="px-4 sm:px-6">
                         <div className="flex items-start justify-between">
                           <Dialog.Title className="text-base font-semibold leading-6 text-light">
-                            View user info
+                            View User Info
                           </Dialog.Title>
                           <div className="ml-3 flex h-7 items-center">
                             <button
@@ -275,7 +299,7 @@ export default function UserActionButton({ selectedUser }: { selectedUser: User 
                           </div>
                         </div>
                       </div>
-                      <div className="relative mt-6 flex-1 flex-wrap overflow-hidden px-4 sm:px-6">
+                      <div className="relative flex-1 flex-wrap overflow-auto">
                         <UserInfo userId={selectedUser.id} />
                       </div>
                     </div>
